refactor!: expose all exported functions in namespaces
Introduce `guiUtils` and `requestUtils` to hold functions that previously lived alone.
By convention, only the global configuration functions live outside a namespace.

BREAKING CHANGES: some functions are now accessible via a namespace:
  - `get`, `getAll`, `load`, `submit` via `requestUtils`
  - `error`, `popup` via `guiUtils`
