+++ packages/core/src/gui/MaxWindow.ts
@@ -23,12 +23,8 @@ import { fit, getCurrentStyle } from '../util/styleUtils';
 import InternalEvent from '../view/event/InternalEvent';
 import Client from '../Client';
 import { NODETYPE } from '../util/Constants';
-import { br, write } from '../util/domUtils';
+import { write } from '../util/domUtils';
 import { getClientX, getClientY } from '../util/EventUtils';
-import { htmlEntities } from '../util/StringUtils';
-import { utils } from '../util/Utils';
-
-import { translate } from '../internal/i18n-utils';
 
 let activeWindow: MaxWindow | null = null;
 
@@ -179,7 +175,7 @@ let activeWindow: MaxWindow | null = null;
  *
  * @category GUI
  */
-class MaxWindow extends EventSource {
+export default class MaxWindow extends EventSource {
   constructor(
     title: string,
     content: HTMLElement | null,
@@ -959,140 +955,3 @@ class MaxWindow extends EventSource {
     this.contentWrapper = null;
   }
 }
-
-/**
- * Shows the specified text content in a new <MaxWindow> or a new browser
- * window if isInternalWindow is false.
- *
- * @param content String that specifies the text to be displayed.
- * @param isInternalWindow Optional boolean indicating if an MaxWindow should be
- * used instead of a new browser window. Default is false.
- */
-export const popup = (content: string, isInternalWindow = false) => {
-  if (isInternalWindow) {
-    const div = document.createElement('div');
-
-    div.style.overflow = 'scroll';
-    div.style.width = '636px';
-    div.style.height = '460px';
-
-    const pre = document.createElement('pre');
-    pre.innerHTML = htmlEntities(content, false)
-      .replace(/\n/g, '<br>')
-      .replace(/ /g, '&nbsp;');
-
-    div.appendChild(pre);
-
-    const w = document.body.clientWidth;
-    const h = Math.max(
-      document.body.clientHeight || 0,
-      document.documentElement.clientHeight
-    );
-    const wnd = new MaxWindow(
-      'Popup Window',
-      div,
-      w / 2 - 320,
-      h / 2 - 240,
-      640,
-      480,
-      false,
-      true
-    );
-
-    wnd.setClosable(true);
-    wnd.setVisible(true);
-  } else {
-    // Wraps up the XML content in a textarea
-    if (Client.IS_NS) {
-      const wnd = window.open();
-      if (!wnd) {
-        throw new Error('Permission not granted to open popup window');
-      }
-      wnd.document.writeln(`<pre>${htmlEntities(content)}</pre`);
-      wnd.document.close();
-    } else {
-      const wnd = window.open();
-      if (!wnd) {
-        throw new Error('Permission not granted to open popup window');
-      }
-      const pre = wnd.document.createElement('pre');
-      pre.innerHTML = htmlEntities(content, false)
-        .replace(/\n/g, '<br>')
-        .replace(/ /g, '&nbsp;');
-      wnd.document.body.appendChild(pre);
-    }
-  }
-};
-
-/**
- * Displays the given error message in a new <MaxWindow> of the given width.
- * If close is true then an additional close button is added to the window.
- * The optional icon specifies the icon to be used for the window. Default
- * is {@link Utils#errorImage}.
- *
- * @param message String specifying the message to be displayed.
- * @param width Integer specifying the width of the window.
- * @param close Optional boolean indicating whether to add a close button.
- * @param icon Optional icon for the window decoration.
- */
-export const error = (
-  message: string,
-  width: number,
-  close: boolean,
-  icon: string | null = null
-) => {
-  const div = document.createElement('div');
-  div.style.padding = '20px';
-
-  const img = document.createElement('img');
-  img.setAttribute('src', icon || utils.errorImage);
-  img.setAttribute('valign', 'bottom');
-  img.style.verticalAlign = 'middle';
-  div.appendChild(img);
-
-  div.appendChild(document.createTextNode('\u00a0')); // &nbsp;
-  div.appendChild(document.createTextNode('\u00a0')); // &nbsp;
-  div.appendChild(document.createTextNode('\u00a0')); // &nbsp;
-  write(div, message);
-
-  const w = document.body.clientWidth;
-  const h = document.body.clientHeight || document.documentElement.clientHeight;
-  const warn = new MaxWindow(
-    translate(utils.errorResource) || utils.errorResource,
-    div,
-    (w - width) / 2,
-    h / 4,
-    width,
-    null,
-    false,
-    true
-  );
-
-  if (close) {
-    br(div);
-
-    const tmp = document.createElement('p');
-    const button = document.createElement('button');
-
-    button.setAttribute('style', 'float:right');
-
-    InternalEvent.addListener(button, 'click', (evt: MouseEvent) => {
-      warn.destroy();
-    });
-
-    write(button, translate(utils.closeResource) || utils.closeResource);
-
-    tmp.appendChild(button);
-    div.appendChild(tmp);
-
-    br(div);
-
-    warn.setClosable(true);
-  }
-
-  warn.setVisible(true);
-
-  return warn;
-};
-
-export default MaxWindow;
